lenguaje español y algunos ajustes

+++ app/Http/Controllers/ProfileController.php
@@ -80,7 +80,9 @@ class ProfileController extends Controller
         $user = User::find($id);
         // Eliminar la imagen, si existe
         if ($user->image && Storage::disk('public')->exists('img/'. $user->image)) {
-            Storage::disk('public')->delete('img/'. $user->image);
+            if($user->image !== 'perfil.png'){
+                Storage::disk('public')->delete('img/' . $user->image);
+            }
         }
 
         // Eliminar el usuario de la base de datos
